requirements.txt
requests>=2.31
beautifulsoup4>=4.12
numpy>=1.24
umap-learn>=0.5
hdbscan>=0.8.33
google-generativeai>=0.5
python-dotenv>=1.0
